Add term feature

+++ src/main/java/com/example/programschemes/model/Batches.java
@@ -45,7 +45,7 @@ public class Batches {
     private String bchinstcode;
 
     @Column(name = "scheme_id")
-    private Long schemeId;
+    private Integer schemeId;
 
     public Short getBchid() { return bchid; }
     public void setBchid(Short bchid) { this.bchid = bchid; }
@@ -83,6 +83,6 @@ public class Batches {
     public String getBchinstcode() { return bchinstcode; }
     public void setBchinstcode(String bchinstcode) { this.bchinstcode = bchinstcode; }
 
-    public Long getSchemeId() { return schemeId; }
-    public void setSchemeId(Long schemeId) { this.schemeId = schemeId; }
+    public Integer getSchemeId() { return schemeId; }
+    public void setSchemeId(Integer schemeId) { this.schemeId = schemeId; }
 }
